Initial check-in for HZN-605: Consolidate TrapdIpMgr implementations
