+++ features/events/traps/src/main/resources/META-INF/opennms/applicationContext-trapDaemon.xml
@@ -24,7 +24,7 @@
     <property name="backlogQ" ref="trapdBacklogQ"/>
   </bean>
   
-  <bean id="trapdIpMgr" class="org.opennms.netmgt.trapd.HibernateTrapdIpMgr"/>
+  <bean id="trapdIpMgr" class="org.opennms.netmgt.trapd.TrapdIpManagerDaoImpl"/>
   
   <bean id="eventReader" class="org.opennms.netmgt.trapd.BroadcastEventProcessor">
     <constructor-arg ref="eventIpcManager" />
