+++ features/events/traps/src/test/resources/org/opennms/netmgt/trapd/trapdIpMgr-test.xml
@@ -19,6 +19,6 @@
   <context:annotation-config />
   <tx:annotation-driven/>
 
-  <bean id="trapdIpMgr" class="org.opennms.netmgt.trapd.HibernateTrapdIpMgr"/>
+  <bean id="trapdIpMgr" class="org.opennms.netmgt.trapd.TrapdIpManagerDaoImpl"/>
 
 </beans>
